# CONNECT-HUB-Api-Gateway/cmd/docs/swagger.yaml
definitions:
...
  models.JobOpening:
    properties:
      application_deadline:
        type: string
      description:
        type: string
      education_level:
        type: string
      employer_id:
        type: integer
      employment_type:
        type: string
      experience_level:
        type: string
      location:
        type: string
      requirements:
        type: string
      salary:
        type: string
      skills_required:
        type: string
      title:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: ConnectHub is a professional networking platform.
  title: Go + Gin professional networking platform API Connect Hub
  version: 1.0.0
paths:
...
  /recruiter/job:
    delete:
      consumes:
      - application/json
      description: Delete a job posted by a recruiter
      parameters:
      - description: ID of the job to delete
        in: query
        name: job_id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job successfully deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Failed to delete job: missing or incorrect parameters'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to delete job'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Delete a job
      tags:
      - Recruiter
    get:
      consumes:
      - application/json
      description: Retrieve the details of a single job posted by a recruiter
      parameters:
      - description: ID of the job to retrieve
        in: query
        name: job_id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved job
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Failed to retrieve job: missing or incorrect parameters'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to retrieve job'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get one job
      tags:
      - Recruiter
    put:
      consumes:
      - application/json
      description: Update a job posted by a recruiter
      parameters:
      - description: ID of the job to update
        in: query
        name: job_id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated job details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JobOpening'
      produces:
      - application/json
      responses:
        "200":
          description: Job successfully updated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Failed to update job: missing or incorrect parameters'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to update job'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update a job
      tags:
      - Recruiter
...
  /recruiter/jobs:
    get:
      consumes:
      - application/json
      description: Retrieve all jobs posted by a recruiter
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved jobs
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Failed to retrieve jobs: missing or incorrect parameters'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to retrieve jobs'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get all jobs
      tags:
      - Recruiter
    post:
      consumes:
      - application/json
      description: Create a new job posting by a recruiter
      parameters:
      - description: Job opening data for posting
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JobOpening'
      produces:
      - application/json
      responses:
        "200":
          description: Job posted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Failed to post job: missing or incorrect parameters'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: failed to post job'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Create job posting
      tags:
      - Recruiter
...
securityDefinitions:
  BearerTokenAuth:
    description: 'For accessing the API, you need to include "Jobseeker","Admin" or
      "Recruiter" before your token in the Authorization header. Example: "Authorization:
      Admin <your_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"